test: move utils tests to spec files
