--- libs/utils/test/object.test.ts
@@ -1,20 +1,0 @@
-import { ColumnType } from 'typeorm'
-import { createEmpty } from '../src'
-
-describe('Object.createEmpty', () => {
-  it('Returns 0 for number', () => {
-    expect(createEmpty(Number)).toEqual(0)
-  })
-  it('Returns empty string for everything else', () => {
-    const tests: ColumnType[] = [
-      String,
-      'longblob',
-      'text',
-      'json',
-      'mediumblob',
-    ]
-    tests.forEach((one) => {
-      expect(createEmpty(one)).toEqual('')
-    })
-  })
-})
